Be explicit about types to avoid rounding errors

+++ du/diskusage.go
@@ -30,7 +30,7 @@ func (du *DiskUsage) Available() uint64 {
 
 // Size returns total size of the file system
 func (du *DiskUsage) Size() uint64 {
-	return du.stat.Blocks * uint64(du.stat.Bsize)
+	return uint64(du.stat.Blocks) * uint64(du.stat.Bsize)
 }
 
 // Used returns total bytes used in file system
